Update main.js 3 CARD AT A TIME UPDATE
This is when I updated from the last working javascript that I know was sending votes to the google sheets. I am notating this so I know if it does not work I can revert it. 

The goal of this is to get only 3 cards to appear at a time.